Aula 1/mnist/requirements.txt
joblib
numpy
scikit-learn
pandas
fastapi
uvicorn
joblib
numpy
pillow
scikit-learn
python-multipart